Delete aquery search
