+++ packages/common/src/api/reducer.ts
@@ -5,7 +5,6 @@ import { authorizedAppsApiReducer } from './authorizedApps'
 import { developerAppsApiReducer } from './developerApps'
 import { favoritesApiReducer } from './favorites'
 import { libraryApiReducer } from './library'
-import { searchApiReducer } from './search'
 import { signUpReducer } from './signUp'
 import { trendingApiReducer } from './trending'
 import { userApiReducer } from './user'
@@ -16,7 +15,6 @@ export default combineReducers({
   developerAppsApi: developerAppsApiReducer,
   favoritesApi: favoritesApiReducer,
   libraryApi: libraryApiReducer,
-  searchApi: searchApiReducer,
   signUpApi: signUpReducer,
   trendingApi: trendingApiReducer,
   userApi: userApiReducer
